+++ Agent/Agent/Agent/Networks/ACER/ReplayMemory.h
@@ -53,7 +53,7 @@ public:
 	*/
 	void StoreStateTransition(
 		State state,
-		vector<float> actions,
+		vector<int> actions,
 		float reward,
 		State newState,
 		bool terminal,
